Do not accept commands if not in SLE ACTIVE state
Modify TcSleLink to avoid accepting commands if SLE is not ACTIVE.
Change the SLE state monitor to enable and disable actions even if
commanding is not available, so long as the link is not disabled.
Modify the start action to execute even if commanding is not available,
so long as the link is not disabled.